vercye_webapp/requirements.txt
celery[redis]
fastapi[standard]
pyyaml
